requirements.txt
biopython
numpy==1.19.3